Further seperate Osmosis to object files
