+++ cpp/Osmosis/ObjectStore/LabelLogAppender.h
@@ -1,8 +1,13 @@
 #ifndef __OSMOSIS_OBJECT_STORE_LABEL_LOG_APPENDER_H__
 #define __OSMOSIS_OBJECT_STORE_LABEL_LOG_APPENDER_H__
 
+#include <string>
+#include <boost/filesystem/path.hpp>
+#include <boost/asio/ip/host_name.hpp>
 #include "Osmosis/ObjectStore/LabelLogEntry.h"
 #include "Osmosis/ObjectStore/MakeDirectory.h"
+#include "Osmosis/ObjectStore/DirectoryNames.h"
+#include "Osmosis/FileDescriptor.h"
 #include "Osmosis/Debug.h"
 
 namespace Osmosis {
@@ -12,48 +17,17 @@ namespace ObjectStore
 class LabelLogAppender
 {
 public:
-	LabelLogAppender( const boost::filesystem::path & rootPath ) :
-		_labelLogsPath( rootPath / DirectoryNames::LABEL_LOG ),
-		_makeDirectory( rootPath / DirectoryNames::LABEL_LOG )
-	{}
+	LabelLogAppender( const boost::filesystem::path & rootPath );
 
-	~LabelLogAppender()
-	{
-		try {
-			write();
-		} CATCH_ALL_IGNORE( "Unable to write label log on object store closing" );
-	}
+	~LabelLogAppender();
 
-	void set(const std::string & label, const Hash & hash)
-	{
-		append(LabelLogEntry::SET, label, hash);
-	}
+	void set(const std::string & label, const Hash & hash);
 
-	void get(const std::string & label, const Hash & hash)
-	{
-		append(LabelLogEntry::GET, label, hash);
-	}
+	void get(const std::string & label, const Hash & hash);
 
-	void remove(const std::string & label, const Hash & hash)
-	{
-		append(LabelLogEntry::REMOVE, label, hash);
-	}
+	void remove(const std::string & label, const Hash & hash);
 
-	void write()
-	{
-		if ( _log.size() == 0 )
-			return;
-		Log temp;
-		{
-			std::lock_guard< std::mutex > lock( _logLock );
-			temp.splice(temp.begin(), _log);
-		}
-		_makeDirectory.makeSureExists();
-		boost::filesystem::path logFilePath = _labelLogsPath / logBasename();
-		std::ofstream output( logFilePath.string() );
-		for ( auto & entry : temp )
-			output << entry;
-	} 
+	void write();
 
 private:
 	enum { MAXIMUM_LOG_TO_KEEP_IN_MEMORY = 100 };
@@ -65,17 +39,7 @@ private:
 	std::mutex               _logLock;
 	MakeDirectory            _makeDirectory;
 
-	void append( LabelLogEntry::Operation operation, const std::string & label, const Hash & hash )
-	{
-		Log temp;
-		temp.emplace_back( operation, label, hash );
-		{
-			std::lock_guard< std::mutex > lock( _logLock );
-			_log.splice(_log.begin(), temp);
-		}
-		if ( _log.size() > MAXIMUM_LOG_TO_KEEP_IN_MEMORY )
-			write();
-	}
+	void append( LabelLogEntry::Operation operation, const std::string & label, const Hash & hash );
 
 	static std::string randomHex()
 	{
