+++ cpp/Osmosis/ObjectStore/LabelLogEntry.h
@@ -2,6 +2,7 @@
 #define __OSMOSIS_OBJECT_STORE_LABEL_LOG_ENTRY_H__
 
 #include <boost/algorithm/string.hpp>
+#include <Osmosis/Hash.h>
 #include "Common/SplitString.h"
 
 namespace Osmosis {
@@ -17,50 +18,11 @@ struct LabelLogEntry
 	std::string              label;
 	std::unique_ptr< Hash >  hash;
 
-	LabelLogEntry( Operation operation, const std::string label, const Hash & hash ) :
-		time( std::time( nullptr ) ),
-		operation( operation ),
-		label( label ),
-		hash( new Hash( hash ) )
-	{}
+	LabelLogEntry( Operation operation, const std::string label, const Hash & hash );
 
-	LabelLogEntry( const LabelLogEntry & other ) :
-		time( other.time ),
-		operation( other.operation ),
-		label( other.label ),
-		hash( new Hash( * other.hash ) )
-	{}
+	LabelLogEntry( const LabelLogEntry & other );
 
-	LabelLogEntry( std::string line )
-	{
-		boost::trim( line );
-		SplitString split( line, '\t' );
-		time = std::stol( split.asString() );
-		split.next();
-		if ( split.done() )
-			THROW( Error, "'" << line << "' is in an invalid format for a label log entry" );
-		operation = parseOperation( split.asString() );
-		split.next();
-		if ( split.done() )
-			THROW( Error, "'" << line << "' is in an invalid format for a label log entry" );
-		label = std::move( split.asString() );
-		split.next();
-		if ( split.done() )
-			return;
-		hash.reset( new Hash( split.asString() ) );
-		split.next();
-		if ( not split.done() )
-			THROW( Error, "'" << line << "' is in an invalid format for a label log entry" );
-	}
-
-	friend std::ostream & operator<<( std::ostream & os, const LabelLogEntry & entry )
-	{
-		os << entry.time << '\t' << static_cast< char >( entry.operation ) << '\t' << entry.label;
-		if ( entry.hash )
-			os << '\t' << * entry.hash;
-		os << '\n';
-		return os;
-	}
+	LabelLogEntry( std::string line );
 
 private:
 	static Operation parseOperation( const std::string & input )
@@ -76,6 +38,8 @@ private:
 
 	LabelLogEntry & operator= ( const LabelLogEntry & rhs ) = delete;
 };
+
+std::ostream & operator<<( std::ostream & os, const LabelLogEntry & entry );
 
 } // namespace ObjectStore
 } // namespace Osmosis
