+++ cpp/Osmosis/Stream/AckOps.h
@@ -1,6 +1,10 @@
 #ifndef __OSMOSIS_STREAM_ACK_OPS_H__
 #define __OSMOSIS_STREAM_ACK_OPS_H__
 
+#include <Common/Error.h>
+#include <Osmosis/TCPSocket.h>
+#include <Osmosis/Stream/AckOps.h>
+
 namespace Osmosis {
 namespace Stream
 {
@@ -8,28 +12,11 @@ namespace Stream
 class AckOps
 {
 public:
-	AckOps( TCPSocket & socket ):
-		_socket( socket )
-	{}
+	AckOps( TCPSocket & socket );
 
-	void wait( const char * waitReason )
-	{
-		try {
-			struct Tongue::Header header = _socket.receiveAll< struct Tongue::Header >();
-			if ( header.opcode != static_cast< unsigned >( Tongue::Opcode::ACK ) )
-				THROW( Error, "Peer did not ack, found byte " << header.opcode << ", when " << waitReason );
-		} catch ( boost::system::system_error & e ) {
-			if ( e.code() == boost::asio::error::eof )
-				THROW( Error, "Peer terminated connection instead of sending ack, when " << waitReason );
-			throw;
-		}
-	}
+	void wait( const char * waitReason );
 
-	void sendAck()
-	{
-		struct Tongue::Header header = { static_cast< unsigned char >( Tongue::Opcode::ACK ) };
-		_socket.sendAll( header );
-	}
+	void sendAck();
 
 private:
 	TCPSocket &            _socket;
