+++ cpp/Osmosis/main.cpp
@@ -6,6 +6,7 @@
 #include "Osmosis/Client/LabelOps.h"
 #include "Osmosis/ObjectStore/LabelLogIterator.h"
 #include "Osmosis/ObjectStore/LeastRecentlyUsed.h"
+#include "Osmosis/ObjectStore/Purge.h"
 
 std::mutex globalTraceLock;
